+++ cloudy_salesforce/sobjects/__init__.py
@@ -1,6 +1,7 @@
 from .sobject import (
     SObjects,
     get_sobject_registry,
+    get_sobject_type_hints,
     parse_sobject_response,
     sobject,
 )
@@ -8,6 +9,7 @@ from .sobject import (
 __all__ = [
     "SObjects",
     "get_sobject_registry",
+    "get_sobject_type_hints",
     "parse_sobject_response",
     "sobject",
 ]